Changes to TRT-LLM download tool for multigpu distributed case
